Further re-arranged the modules to make the system more decoupled.
Implemented a mechanism to be able to dynamically deploy rest-api endpoints via the event bus.

Added slf4j and log4j2 logging.

Added batch scripts in the tools module to provide ways for the system to be easily run.

Added readme-s in the modules for better description of the functionality.